+++ src/pages/App.tsx
@@ -2,6 +2,7 @@ import React, { useEffect } from 'react';
 import { Helmet } from 'react-helmet';
 import { Navigate, Outlet } from 'react-router-dom';
 
+import SwapPage from '@src/pages/SwapPage';
 import ErrorBoundary from '@src/components/general/ErrorBoundry';
 import NavigationBar from '@src/components/nugg/PageLayout/NavigationBar/NavigationBar';
 import ToastContainer from '@src/components/general/Toast/ToastContainer';
@@ -9,11 +10,10 @@ import Modal from '@src/components/nugg/Modals/Modal/Modal';
 import { useRoutes } from '@src/lib/router';
 import client from '@src/client';
 import useAnalyticsReporter from '@src/lib/analytics/useAnalyticsReporter';
-import SwapPage from '@src/structure/desktop/SwapPage';
 
-const WalletView = React.lazy(() => import('@src/structure/mobile/WalletView'));
-const HotRotateO = React.lazy(() => import('@src/structure/desktop/HotRotateO'));
-const SearchOverlay = React.lazy(() => import('@src/structure/desktop/SearchOverlay'));
+const MobileWalletView = React.lazy(() => import('@src/pages/mobile/MobileWalletView'));
+const HotRotateO = React.lazy(() => import('@src/pages/HotRotateO'));
+const SearchOverlay = React.lazy(() => import('@src/pages/SearchOverlay'));
 
 const App = () => {
     useAnalyticsReporter();
@@ -34,8 +34,9 @@ const App = () => {
             element: <Outlet />,
             children: [
                 { path: 'view/*', element: loaded ? <SearchOverlay /> : null, overlay: 997 },
-                { path: 'edit', element: <HotRotateO /> },
-                { path: 'wallet', element: <WalletView /> },
+                { path: 'edit/:id', element: <HotRotateO /> },
+                // instead of hiding this for mobile here, we redirect inside the component to avoid lots of rerenders
+                { path: 'wallet', element: <MobileWalletView /> },
                 { path: 'swap/:id', element: null },
                 { path: '*', element: <Navigate to={`swap/${epoch || ''}`} /> },
             ],
