--- src/structure/desktop/SearchOverlay.styles.ts
@@ -1,29 +1,0 @@
-import { NLStyleSheetCreator } from '@src/lib';
-
-const styles = NLStyleSheetCreator({
-    container: {
-        display: 'flex',
-        justifyContent: 'space-between',
-        alignItems: 'space-between',
-    },
-    nuggDexContainer: {
-        display: 'flex',
-        width: '40%',
-        height: '100%',
-        flexDirection: 'column',
-        justifyContent: 'center',
-        alignItems: 'center',
-    },
-    tokenContainer: {
-        display: 'flex',
-        width: '40%',
-        height: '100%',
-        flexDirection: 'column',
-        justifyContent: 'center',
-        alignItems: 'center',
-        position: 'relative',
-        overflow: 'scroll',
-    },
-});
-
-export default styles;
