--- src/structure/mobile/SwapView/SwapView.styles.ts
@@ -1,31 +1,0 @@
-import { NLStyleSheetCreator } from '@src/lib';
-
-const styles = NLStyleSheetCreator({
-    container: {
-        display: 'flex',
-        flexDirection: 'column',
-        justifyContent: 'flex-start',
-        alignItems: 'center',
-        height: '100%',
-        width: '100%',
-    },
-    ring: {
-        width: '100%',
-        height: '70%',
-        display: 'flex',
-        flexDirection: 'column',
-        justifyContent: 'center',
-        alignItems: 'center',
-        position: 'absolute',
-    },
-    ringAbout: {
-        width: '100%',
-        height: '100%',
-        display: 'flex',
-        flexDirection: 'column',
-        justifyContent: 'flex-end',
-        alignItems: 'center',
-    },
-});
-
-export default styles;
