--- src/structure/mobile/index.tsx
@@ -1,9 +1,0 @@
-import { FunctionComponent } from 'react';
-
-type Props = Record<string, never>;
-
-const Mobile: FunctionComponent<Props> = () => {
-    return null;
-};
-
-export default Mobile;
